<test>(domain/NameTest.java): 테스트 추가
